Revert "update"
This reverts commit 03ec3d072a176872476f2c3877e4b1220bc0c490.

+++ yolov8nd/models/yolo/detect/train.py
@@ -22,9 +22,9 @@ class DetectionTrainer(BaseTrainer):
 
     Example:
         ```python
-        from ultralytics.models.yolo.detect import DetectionTrainer
+        from yolov8nd.models.yolo.detect import DetectionTrainer
 
-        args = dict(model="yolov8n.pt", data="coco8.yaml", epochs=3)
+        args = dict(model='yolov8n.pt', data='coco8.yaml', epochs=3)
         trainer = DetectionTrainer(overrides=args)
         trainer.train()
         ```
@@ -60,7 +60,7 @@ class DetectionTrainer(BaseTrainer):
         if self.args.multi_scale:
             imgs = batch["img"]
             sz = (
-                random.randrange(int(self.args.imgsz * 0.5), int(self.args.imgsz * 1.5 + self.stride))
+                random.randrange(self.args.imgsz * 0.5, self.args.imgsz * 1.5 + self.stride)
                 // self.stride
                 * self.stride
             )  # size
